Support for summaries per api method
Because the summary ops are created within a branch of the graph
corresponding to an api method invocation, attempt to evaluate all
of the summary operations upon api execution results in error.

This change tracks the summary ops per api method and so can
safely evaluate only the necessary summary ops. These summaries
are tracked on training step timeline.

As part of this change non-api summaries are also tracked: the
reward per step and reward per episode on environment interactions
timeline.

As an example the PPOAgent is instrumented to track the policy and
the value function losses.

TODOs:
- Tensorflow eager
- Pytorch
- respect the summary spec regex
- properly form summary ops names including namespaces
- in multi-environment setup track the environment when reporting
  rewards